+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * NP-756 Advisernet related content component
+* NP-472 New OISC Warning component
 
 ## <sub>v1.10.1</sub>
 
@@ -6,7 +7,7 @@
 
 * NP-924 Callout headings bug
 * NP-895 Correct footer class and SCSS
-* NP-895 Fix search appearance 
+* NP-895 Fix search appearance
 
 ## <sub>v1.10.0</sub>